days/day-098-integration-project/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
